move profile plot figure to plot_functions.py
