+++ app/plot_functions.py
@@ -7,6 +7,7 @@ from typing import Literal
 import pandas as pd
 import plotly.express as px
 import plotly.graph_objects as go
+import pypsa
 import streamlit as st
 
 from app.ptxboa_functions import (
@@ -485,3 +486,207 @@ def create_scatter_plot(df_res, settings: dict):
     fig.update_traces(texttemplate="%{text}", textposition="top center")
     st.plotly_chart(fig)
     st.write(df_res)
+
+
+def create_profile_figure(n: pypsa.Network) -> None:
+    def transform_time_series(
+        df: pd.DataFrame, parameter: str = "Power"
+    ) -> pd.DataFrame:
+        res = df.reset_index().melt(
+            id_vars=["timestep", "period"],
+            var_name="Component",
+            value_name="MW (MWh for SOC)",
+        )
+        res["Parameter"] = parameter
+        return res
+
+    df_p_max_pu = n.generators_t["p_max_pu"]
+    df_p_max_pu = transform_time_series(df_p_max_pu, parameter="cap. factor")
+    df_gen = n.generators_t["p"]
+    df_gen = transform_time_series(df_gen)
+    df_links = -n.links_t["p1"]
+    df_links = transform_time_series(df_links)
+    df_store = n.stores_t["e"]
+    df_store = transform_time_series(df_store)
+    df_storageunit = n.storage_units_t["state_of_charge"]
+    df_storageunit = transform_time_series(df_storageunit)
+
+    df = pd.concat([df_p_max_pu, df_gen, df_links, df_store, df_storageunit])
+
+    # selection:
+    df = df.loc[
+        df["Component"].isin(
+            [
+                "PV-FIX",
+                "WIND-ON",
+                "WIND-OFF",
+                "ELY",
+                "DERIV",
+                "H2_STR_store",
+                "EL_STR",
+                "final_product_storage",
+            ]
+        )
+    ]
+
+    # rename components:
+    rename_list = {
+        "PV-FIX": "PV tilted",
+        "WIND-ON": "Wind onshore",
+        "WIND-OFF": "Wind offshore",
+        "ELY": "Electrolyzer",
+        "DERIV": "Derivate production",
+        "H2_STR_in": "H2 storage",
+        "H2_STR_store": "H2 storage",
+        "final_product_storage": "Final product storage",
+        "EL_STR": "Electricity storage",
+        "CO2-G_supply": "CO2 supply",
+        "H2O-L_supply": "Water supply",
+    }
+    df = df.replace(rename_list)
+
+    df_sel = df
+
+    # add continous time index:
+    df_sel["period"] = df_sel["period"].astype(int)
+    df_sel["timestep"] = df_sel["timestep"].astype(int)
+    df_sel["time"] = 7 * 24 * df_sel["period"] + df_sel["timestep"]
+    df_sel = df_sel.sort_values("time")
+
+    # generation:
+    st.subheader("Output")
+    fig = go.Figure()
+
+    def add_vertical_lines(fig: go.Figure):
+        """Add vertical lines between periods."""
+        for x in range(7 * 24, 7 * 8 * 24, 7 * 24):
+            fig.add_vline(x=x, line_color="black", line_width=0.5)
+
+    def add_to_figure(
+        df: pd.DataFrame,
+        fig: go.Figure,
+        component: str,
+        parameter: str,
+        color: str,
+        fill: bool = False,
+    ):
+        df_plot = df[(df["Component"] == component)]
+        df_plot = df_plot[(df_plot["Parameter"] == parameter)]
+        if fill:
+            fig.add_trace(
+                go.Line(
+                    x=df_plot["time"],
+                    y=df_plot["MW (MWh for SOC)"],
+                    name=component,
+                    line_color=color,
+                    stackgroup="one",
+                )
+            )
+        else:
+            fig.add_trace(
+                go.Line(
+                    x=df_plot["time"],
+                    y=df_plot["MW (MWh for SOC)"],
+                    name=component,
+                    line_color=color,
+                )
+            )
+
+    add_to_figure(
+        df_sel, fig, component="PV tilted", parameter="Power", fill=True, color="yellow"
+    )
+    add_to_figure(
+        df_sel,
+        fig,
+        component="Wind onshore",
+        parameter="Power",
+        fill=True,
+        color="blue",
+    )
+    add_to_figure(
+        df_sel,
+        fig,
+        component="Wind offshore",
+        parameter="Power",
+        fill=True,
+        color="blue",
+    )
+    add_to_figure(
+        df_sel, fig, component="Electrolyzer", parameter="Power", color="black"
+    )
+    add_to_figure(
+        df_sel, fig, component="Derivate production", parameter="Power", color="red"
+    )
+
+    add_vertical_lines(fig)
+
+    fig.update_layout(
+        xaxis={"title": "time (h)"},
+        yaxis={"title": "output (MW)"},
+    )
+
+    st.plotly_chart(fig, use_container_width=True)
+
+    st.subheader("Storage state of charge")
+    include_final_product_storage = st.toggle("Show final product storage", value=False)
+    # storage figure:
+    fig = go.Figure()
+
+    add_to_figure(
+        df_sel,
+        fig,
+        component="Electricity storage",
+        parameter="Power",
+        color="black",
+    )
+
+    add_to_figure(
+        df_sel,
+        fig,
+        component="H2 storage",
+        parameter="Power",
+        color="red",
+    )
+    if include_final_product_storage:
+        add_to_figure(
+            df_sel,
+            fig,
+            component="Final product storage",
+            parameter="Power",
+            color="blue",
+        )
+
+    add_vertical_lines(fig)
+
+    fig.update_layout(
+        xaxis={"title": "time (h)"},
+        yaxis={"title": "state of charge (MWh)"},
+    )
+
+    st.plotly_chart(fig, use_container_width=True)
+
+    st.subheader("Capacity factors")
+
+    fig = go.Figure()
+    add_to_figure(
+        df_sel, fig, component="PV tilted", parameter="cap. factor", color="yellow"
+    )
+    add_to_figure(
+        df_sel,
+        fig,
+        component="Wind onshore",
+        parameter="cap. factor",
+        color="blue",
+    )
+    add_to_figure(
+        df_sel,
+        fig,
+        component="Wind offshore",
+        parameter="cap. factor",
+        color="blue",
+    )
+    add_vertical_lines(fig)
+    st.plotly_chart(fig, use_container_width=True)
+
+    st.subheader("Profile data")
+    st.dataframe(df_sel, use_container_width=True)
